chore: moved the engine definitions into its own file
